+++ web/app/components/style-guide/loading-bar.component.jsx
@@ -1,0 +1,13 @@
+var React = require('react/addons');
+var StyleGuideTemplate = require('./assets/components/style-guide-template.component.jsx');
+var data = require('./data/loading-bar.jsx');
+
+var LoadingBarPage = React.createClass({
+  render: function() {
+    return (
+      <StyleGuideTemplate data={data} />
+    );
+  }
+});
+
+module.exports = LoadingBarPage;